Run Sample Apps e2e tests

+++ e2e/runner/run_cypress_local.ts
@@ -7,7 +7,7 @@ import {
   shell,
   unBooleanify,
 } from "./cypress-runner-utils";
-import { run as runSampleAppsForEmbeddingSdk } from "./run-sample-apps-for-embedding-sdk/run";
+import { run as runSampleAppForEmbeddingSdk } from "./run-sample-app-for-embedding-sdk/run";
 
 // if you want to change these, set them as environment variables in your shell
 const userOptions = {
@@ -28,6 +28,7 @@ const derivedOptions = {
   CYPRESS_ALL_FEATURES_TOKEN: userOptions.ENTERPRISE_TOKEN,
   QA_DB_ENABLED: userOptions.START_CONTAINERS,
   BUILD_JAR: userOptions.BACKEND_PORT === 4000,
+  START_BACKEND: userOptions.BACKEND_PORT === 4000,
   CYPRESS_IS_EMBEDDING_SDK: userOptions.TEST_SUITE === "component",
   MB_SNOWPLOW_AVAILABLE: userOptions.START_CONTAINERS,
   MB_SNOWPLOW_URL: "http://localhost:9090",
@@ -56,6 +57,7 @@ printBold(`Running Cypress with options:
   - BUILD_JAR          : ${options.BUILD_JAR}
   - GENERATE_SNAPSHOTS : ${options.GENERATE_SNAPSHOTS}
   - BACKEND_PORT       : ${options.BACKEND_PORT}
+  - START_BACKEND      : ${options.START_BACKEND}
   - OPEN_UI            : ${options.OPEN_UI}
   - SHOW_BACKEND_LOGS  : ${options.SHOW_BACKEND_LOGS}
 `);
@@ -70,20 +72,22 @@ const init = async () => {
     printBold("⏳ Building backend");
     shell("./bin/build-for-test");
 
-    const isBackendRunning = shell(
-      `lsof -ti:${options.BACKEND_PORT} || echo ""`,
-      { quiet: true },
-    );
-    if (isBackendRunning) {
-      printBold(
-        "⚠️ Your backend is already running, you may want to kill pid " +
-          isBackendRunning,
+    if (options.START_BACKEND) {
+      const isBackendRunning = shell(
+        `lsof -ti:${options.BACKEND_PORT} || echo ""`,
+        { quiet: true },
       );
-      process.exit(FAILURE_EXIT_CODE);
-    }
+      if (isBackendRunning) {
+        printBold(
+          "⚠️ Your backend is already running, you may want to kill pid " +
+            isBackendRunning,
+        );
+        process.exit(FAILURE_EXIT_CODE);
+      }
 
-    printBold("⏳ Starting backend");
-    await CypressBackend.start();
+      printBold("⏳ Starting backend");
+      await CypressBackend.start();
+    }
   } else {
     printBold(
       `Not building a jar, expecting metabase to be running on port ${options.BACKEND_PORT}. Make sure your metabase instance is running with an h2 app db and the following environment variables:
@@ -98,7 +102,7 @@ const init = async () => {
     shell("rm -f e2e/support/cypress_sample_instance_data.json");
 
     printBold("⏳ Generating snapshots");
-    await runCypress("snapshot", cleanup);
+    await runCypress("snapshot", { exitFunction: cleanup });
   }
 
   const isFrontendRunning = shell("lsof -ti:8080 || echo ''", { quiet: true });
@@ -109,8 +113,19 @@ const init = async () => {
   }
 
   switch (options.TEST_SUITE) {
-    case "sample-apps-embedding-sdk-e2e":
-      await runSampleAppsForEmbeddingSdk();
+    case "metabase-nodejs-react-sdk-embedding-sample-e2e":
+      await runSampleAppForEmbeddingSdk(
+        "metabase-nodejs-react-sdk-embedding-sample",
+      );
+      break;
+
+    case "metabase-nextjs-sdk-embedding-sample-app-router-e2e":
+    case "metabase-nextjs-sdk-embedding-sample-pages-router-e2e":
+      await runSampleAppForEmbeddingSdk("metabase-nextjs-sdk-embedding-sample");
+      break;
+
+    case "shoppy-e2e":
+      await runSampleAppForEmbeddingSdk("shoppy");
       break;
   }
